Move EPSFModel to models module
